+++ src/Endpoint/Endpoints/LocalInventory.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Onetoweb\GoogleMerchant\Endpoint\Endpoints;
+
+use Onetoweb\GoogleMerchant\Endpoint\AbstractEndpoint;
+
+/**
+ * LocalInventory Endpoint.
+ */
+class LocalInventory extends AbstractEndpoint
+{
+    /**
+     * @param string $offerId
+     * @param array $query = []
+     * 
+     * @return array|null
+     */
+    public function list(string $offerId, array $query = [])
+    {
+        return $this->client->get("/inventories/{$this->client->getVersion()}/accounts/{$this->client->getAccountId()}/products/$offerId/localInventories", $query);
+    }
+}